requirements.txt
flask
flask_login
flask_openid_stateless
flask_sqlalchemy
flask_wtf
steam
-e .
